Slouching Detector
This code checks various iterations of the slouching detector. Each Iteration has two parts to the code the first part is for taking a reference positions which are your good posture positions. The second block of code finally calculates from the deviation of the calculated angle if you are slouching are not.